requirements.txt
streamlit
python-constraint
Faker
matplotlib
pandas
